Fix SQLite3 keyword comparison in WHERE clause
SQLite3 requires single-quotes for the comparison to work correctly. Double-quotes will cause SQLite3 to do an additional string comparison with the parameter, such as username=username and password=password. This evaluates to TRUE and caused a bug where participants could get the flag for SQLite 1, 2 and 3 without even doing an SQL injection. The bug was found once the competition had started and was left there to be fair with everyone. This would have been different if it compromised the whole track.