Changed the permissions of the source files to www-data instead of root so that apache can access them (#34)
Not sure if it is an error but when building the dockerfile and running it on a linux machine, the web page sends out an error that no file can be accessed in the pasta directory. looking at the files inside the docker the permissions were set to root:root just as prescribed by the Dockerfile. Apache runs with the www-data user and so i feel like it would make alot more sense if the permissions would be set to www-data:www-data on the files that apache is supposed to access.